migration: consistently use enum for mydumper tool
