+++ aiven_mysql_migrate/enums.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class MySQLMigrateTool(str, Enum):
+    mysqldump = "mysqldump"
+    mydumper = "mydumper"
+
+
+class MySQLMigrateMethod(str, Enum):
+    dump = "dump"
+    replication = "replication"
